#24 Add FileManager interface with SftpFileManager and MockFileManager for mock profile support
